Rename `profiler` folder to `rule_based_profiler`
