+++ great_expectations/rule_based_profiler/domain_builder/column_domain_builder.py
@@ -5,8 +5,10 @@ from great_expectations.core.domain_types import (
     MetricDomainTypes,
     StructuredDomainTypes,
 )
-from great_expectations.profiler.domain_builder.domain import Domain
-from great_expectations.profiler.domain_builder.domain_builder import DomainBuilder
+from great_expectations.rule_based_profiler.domain_builder.domain import Domain
+from great_expectations.rule_based_profiler.domain_builder.domain_builder import (
+    DomainBuilder,
+)
 from great_expectations.validator.validation_graph import MetricConfiguration
 from great_expectations.validator.validator import Validator
