+++ great_expectations/rule_based_profiler/domain_builder/simple_semantic_type_domain_builder.py
@@ -3,11 +3,11 @@ from typing import Any, Dict, List, Optional, Union, cast
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.domain_types import MetricDomainTypes, SemanticDomainTypes
 from great_expectations.profile.base import ProfilerTypeMapping
-from great_expectations.profiler.domain_builder.column_domain_builder import (
+from great_expectations.rule_based_profiler.domain_builder.column_domain_builder import (
     ColumnDomainBuilder,
 )
-from great_expectations.profiler.domain_builder.domain import Domain
-from great_expectations.profiler.domain_builder.inferred_semantic_domain_type import (
+from great_expectations.rule_based_profiler.domain_builder.domain import Domain
+from great_expectations.rule_based_profiler.domain_builder.inferred_semantic_domain_type import (
     InferredSemanticDomainType,
 )
 from great_expectations.validator.validator import MetricConfiguration, Validator
@@ -122,52 +122,40 @@ class SimpleSemanticTypeColumnDomainBuilder(ColumnDomainBuilder):
 
         semantic_column_type: SemanticDomainTypes
         if column_type in (
-            set([type_name.upper() for type_name in ProfilerTypeMapping.INT_TYPE_NAMES])
-            | set(
-                [
-                    type_name.upper()
-                    for type_name in ProfilerTypeMapping.FLOAT_TYPE_NAMES
-                ]
-            )
+            {type_name.upper() for type_name in ProfilerTypeMapping.INT_TYPE_NAMES}
+            | {type_name.upper() for type_name in ProfilerTypeMapping.FLOAT_TYPE_NAMES}
         ):
             semantic_column_type = SemanticDomainTypes.NUMERIC
-        elif column_type in set(
-            [type_name.upper() for type_name in ProfilerTypeMapping.STRING_TYPE_NAMES]
-        ):
+        elif column_type in {
+            type_name.upper() for type_name in ProfilerTypeMapping.STRING_TYPE_NAMES
+        }:
             semantic_column_type = SemanticDomainTypes.TEXT
-        elif column_type in set(
-            [type_name.upper() for type_name in ProfilerTypeMapping.BOOLEAN_TYPE_NAMES]
-        ):
+        elif column_type in {
+            type_name.upper() for type_name in ProfilerTypeMapping.BOOLEAN_TYPE_NAMES
+        }:
             semantic_column_type = SemanticDomainTypes.LOGIC
-        elif column_type in set(
-            [type_name.upper() for type_name in ProfilerTypeMapping.DATETIME_TYPE_NAMES]
-        ):
+        elif column_type in {
+            type_name.upper() for type_name in ProfilerTypeMapping.DATETIME_TYPE_NAMES
+        }:
             semantic_column_type = SemanticDomainTypes.DATETIME
-        elif column_type in set(
-            [type_name.upper() for type_name in ProfilerTypeMapping.BINARY_TYPE_NAMES]
-        ):
+        elif column_type in {
+            type_name.upper() for type_name in ProfilerTypeMapping.BINARY_TYPE_NAMES
+        }:
             semantic_column_type = SemanticDomainTypes.BINARY
-        elif column_type in set(
-            [type_name.upper() for type_name in ProfilerTypeMapping.CURRENCY_TYPE_NAMES]
-        ):
+        elif column_type in {
+            type_name.upper() for type_name in ProfilerTypeMapping.CURRENCY_TYPE_NAMES
+        }:
             semantic_column_type = SemanticDomainTypes.CURRENCY
-        elif column_type in set(
-            [type_name.upper() for type_name in ProfilerTypeMapping.IDENTITY_TYPE_NAMES]
-        ):
+        elif column_type in {
+            type_name.upper() for type_name in ProfilerTypeMapping.IDENTITY_TYPE_NAMES
+        }:
             semantic_column_type = SemanticDomainTypes.IDENTITY
         elif column_type in (
-            set(
-                [
-                    type_name.upper()
-                    for type_name in ProfilerTypeMapping.MISCELLANEOUS_TYPE_NAMES
-                ]
-            )
-            | set(
-                [
-                    type_name.upper()
-                    for type_name in ProfilerTypeMapping.RECORD_TYPE_NAMES
-                ]
-            )
+            {
+                type_name.upper()
+                for type_name in ProfilerTypeMapping.MISCELLANEOUS_TYPE_NAMES
+            }
+            | {type_name.upper() for type_name in ProfilerTypeMapping.RECORD_TYPE_NAMES}
         ):
             semantic_column_type = SemanticDomainTypes.MISCELLANEOUS
         else:
