+++ great_expectations/rule_based_profiler/expectation_configuration_builder/expectation_configuration_builder.py
@@ -2,8 +2,8 @@ from abc import ABC, abstractmethod
 from typing import Dict, Optional
 
 from great_expectations.core.expectation_configuration import ExpectationConfiguration
-from great_expectations.profiler.domain_builder.domain import Domain
-from great_expectations.profiler.parameter_builder.parameter_container import (
+from great_expectations.rule_based_profiler.domain_builder.domain import Domain
+from great_expectations.rule_based_profiler.parameter_builder.parameter_container import (
     ParameterContainer,
 )
