+++ great_expectations/rule_based_profiler/parameter_builder/multi_batch_parameter_builder.py
@@ -4,10 +4,10 @@ from typing import Dict, List, Optional
 import great_expectations.exceptions as ge_exceptions
 from great_expectations import DataContext
 from great_expectations.core.batch import BatchRequest
-from great_expectations.profiler.parameter_builder.parameter_builder import (
+from great_expectations.rule_based_profiler.parameter_builder.parameter_builder import (
     ParameterBuilder,
 )
-from great_expectations.profiler.util import get_batch_ids
+from great_expectations.rule_based_profiler.util import get_batch_ids
 
 
 # TODO: <Alex>ALEX -- If ParameterBuilder already extends ABC, why does this class need to do the same?</Alex>
