+++ tests/rule_based_profiler/data/alice_columnar_table_single_batch_data.csv
