+++ tests/rule_based_profiler/test_rule.py
@@ -3,7 +3,7 @@ from typing import Any
 import pytest
 
 import great_expectations.exceptions as ge_exceptions
-from great_expectations.profiler.parameter_builder.parameter_container import (
+from great_expectations.rule_based_profiler.parameter_builder.parameter_container import (
     get_parameter_value,
 )
